server/requirements.txt
django
django-rest-framework
django-cors-headers
django-filter
psycopg2
